+++ Language/Haskell/TH/Desugar/Core.hs
@@ -1386,3 +1386,15 @@ strictToBang Unpacked  = Bang SourceUnpack SourceStrict
 strictToBang :: Bang -> Bang
 strictToBang = id
 #endif
+
+-- | Convert a 'DType' to a 'DPred'
+dTypeToDPred :: Monad q => DType -> q DPred
+dTypeToDPred (DForallT _ _ _) = impossible "Forall-type used as constraint"
+dTypeToDPred (DAppT t1 t2)   = DAppPr <$> dTypeToDPred t1 <*> pure t2
+dTypeToDPred (DSigT ty ki)   = DSigPr <$> dTypeToDPred ty <*> pure ki
+dTypeToDPred (DVarT n)       = return $ DVarPr n
+dTypeToDPred (DConT n)       = return $ DConPr n
+dTypeToDPred DArrowT         = impossible "Arrow used as head of constraint"
+dTypeToDPred (DLitT _)       = impossible "Type literal used as head of constraint"
+dTypeToDPred DWildCardT      = return DWildCardPr
+dTypeToDPred DStarT          = impossible "Star used as head of constraint"